+++ pkg/application/cluster_app_test.go
@@ -12,15 +12,11 @@ func TestClusterAppDefaults(t *testing.T) {
 		t.Errorf("Cluster name not as expected. Expected %s, Actual: %s", clusterName, cluster.Name)
 	}
 
-	if cluster.ClusterApp.Namespace != cluster.Namespace {
-		t.Errorf("ClusterApp namespace not as expected. Expected %s, Actual: %s", cluster.Namespace, cluster.ClusterApp.Namespace)
+	if cluster.ClusterApp.Organization.GetNamespace() != cluster.GetNamespace() {
+		t.Errorf("ClusterApp namespace not as expected. Expected %s, Actual: %s", cluster.GetNamespace(), cluster.ClusterApp.Organization.GetNamespace())
 	}
 
-	if cluster.DefaultAppsApp.Namespace != cluster.Namespace {
-		t.Errorf("DefaultAppsApp namespace not as expected. Expected %s, Actual: %s", cluster.Namespace, cluster.DefaultAppsApp.Namespace)
-	}
-
-	if cluster.Namespace != cluster.Organization.GetNamespace() {
-		t.Errorf("Cluster namespace doesn't match Org namespace. Expected: %s, Actual: %s", cluster.Organization.GetNamespace(), cluster.Namespace)
+	if cluster.DefaultAppsApp.Organization.GetNamespace() != cluster.GetNamespace() {
+		t.Errorf("DefaultAppsApp namespace not as expected. Expected %s, Actual: %s", cluster.GetNamespace(), cluster.DefaultAppsApp.Organization.GetNamespace())
 	}
 }